+++ src/parnum.h
@@ -20,7 +20,7 @@ char *__strdup(const char *s);
 
 /* Parse a single integer value from a string.
  * It searches in the string to find a valid integer */
-int parnum_parse_int(const char *s);
+long parnum_parse_int(const char *s);
 
 /* Parse a single float value from a string.
  * It searches in the string to find a valid floating point number */
